Fix: stacktrace not shown in kibana
When logging exceptions with logging.exception(), the cf-python-logging
library will emit the stacktrace as a separate field in the JSON. This field
is now being filtered by the BTP and no longer visible in Kibana.

This commit adds the stacktrace to the main 'msg' field and thus makes it
possible again to debug exceptions in our Python applications.